requirements.txt
# Core LLM / API
openai>=1.0.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0

# Retrieval / vector search
pinecone>=5.0.0
fastembed>=0.3.0

# Fine-tuning / modeling
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.28.0
sentencepiece>=0.2.0
datasets>=2.18.0
evaluate>=0.4.0

# Data / evaluation
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
tqdm>=4.66.0

# Config / utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

groq>=0.9.0
